Extracted getSkeletonTypeFromLink to shared utils for reuse
